+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.IntegrationTests/Infrastructure/RequestCalculationResult/AggregatedTimeSeriesQueriesTests.cs
@@ -12,176 +12,645 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Globalization;
 using AutoFixture;
 using Energinet.DataHub.Core.TestCommon;
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResults;
-using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.DeltaTableConstants;
 using Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.EnergyResults;
-using Energinet.DataHub.Wholesale.Common.Interfaces.Models;
 using FluentAssertions;
 using FluentAssertions.Execution;
 using NodaTime;
 using Xunit;
-using Period = Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.EnergyResults.Period;
 
 namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Infrastructure.RequestCalculationResult;
 
-public class AggregatedTimeSeriesQueriesTests : TestBase<AggregatedTimeSeriesQueries>, IClassFixture<DatabricksSqlStatementApiFixture>
+public sealed class AggregatedTimeSeriesQueriesTests : TestBase<AggregatedTimeSeriesQueries>,
+    IClassFixture<DatabricksSqlStatementApiFixture>
 {
-    private const string FirstQuantity = "1.111";
-    private const string FirstQuantityFirstCorrection = "1.222";
-    private const string FirstQuantitySecondCorrection = "1.333";
-    private const string FirstQuantityThirdCorrection = "1.444";
-    private const string SecondQuantity = "2.222";
-    private const string SecondQuantityFirstCorrection = "2.333";
-    private const string SecondQuantitySecondCorrection = "2.444";
-    private const string SecondQuantityThirdCorrection = "2.555";
-    private const string ThirdQuantity = "3.333";
-    private const string FourthQuantity = "4.444";
-    private const string FourthQuantityThirdCorrection = "4.555";
-    private const string GridAreaCodeA = "101";
-    private const string GridAreaCodeB = "201";
-    private const string GridAreaCodeC = "301";
-    private static readonly Guid _firstCalculationId = Guid.Parse("019703e7-98ee-45c1-b343-0cbf185a47d9");
-    private static readonly Guid _secondCalculationId = Guid.Parse("119703e7-98ee-45c1-b343-0cbf185a47d9");
-    private readonly DatabricksSqlStatementApiFixture _fixture;
+    private readonly AggregatedTimeSeriesQueriesData _aggregatedTimeSeriesQueriesData;
 
     public AggregatedTimeSeriesQueriesTests(DatabricksSqlStatementApiFixture fixture)
     {
-        _fixture = fixture;
-        Fixture.Inject(_fixture.DatabricksSchemaManager.DeltaTableOptions);
-        Fixture.Inject(_fixture.GetDatabricksExecutor());
+        Fixture.Inject(fixture.DatabricksSchemaManager.DeltaTableOptions);
+        Fixture.Inject(fixture.GetDatabricksExecutor());
+        _aggregatedTimeSeriesQueriesData = new AggregatedTimeSeriesQueriesData(fixture);
     }
 
+    /*
+2 Test purpose
+══════════════
+
+  The purpose of the tests in this class is to ensure that the SQL queries made by `AggregatedTimeSeriesQueryStatement'
+  using `AggregatedTimeSeriesQueryParameters' are resulting in the expected and correct data. The structure of the
+  queries is not considered, neither is performance. The tests focus solely on whether the returned data correlates with
+  our expectations for a particular query parameter.
+
+
+2.1 Test structure
+──────────────────
+
+  The tests in this file all follow the same pattern:
+  1. We generate a specific query parameter
+  2. We execute the generated query using the sut
+  3. We validate that the expected data is returned
+
+
+  In particular we change the query parameter fields listed below
+  ━━━━━━━━━━━━━━━━━━━━━━━━
+   Query parameter field
+  ────────────────────────
+   Time series type
+   Grid area
+   Energy supplier id
+   Balance responsible id
+   Calculation type
+   Start of period
+   End of period
+  ━━━━━━━━━━━━━━━━━━━━━━━━
+
+  The tests are for the most part grouped in two
+  1. We consider only one time series type
+  2. We consider multiple time series types
+
+
+2.2 Result validation
+─────────────────────
+
+  For the case with only one time series type, we, for the most part, consider
+  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+   Type                      Validation
+  ──────────────────────────────────────────────
+   Aggregated time series    version, grid area
+   Energy time series point  quantity
+  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+  For the case with multiple time series types, we limit ourselves to only validating the time series, as the point
+  validating is achieved in the former case
+  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+   Type                      Validation
+  ────────────────────────────────────────────────────────────────
+   Aggregated time series    version, grid area, time series type
+   Energy time series point
+  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+     */
+
     [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorTotalProduction_ReturnsResult()
+    public async Task
+        GetAsync_WhenRequestFromEnergySupplierWithSpecificBalanceResponsibleAndGridArea_CorrectTimeSeriesPoints()
     {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeB,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierB,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleB);
 
         // Act
         var actual = await Sut.GetAsync(parameters).ToListAsync();
 
-        // Assert
         using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
+        actual.Select(ats => ats.Version).Should().BeEquivalentTo([256L, 512L, 1024L, 2048L]);
+
+        actual.Single(ats => ats.Version == 256L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
             .Should()
-            .Equal(FirstQuantity, ThirdQuantity, FourthQuantity);
+            .BeEquivalentTo([2.222, 3.333]);
+
+        actual.Single(ats => ats.Version == 512L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.333]);
+
+        actual.Single(ats => ats.Version == 1024L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.444, 3.555]);
+
+        actual.Single(ats => ats.Version == 2048L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([3.666]);
+    }
+
+    [Fact]
+    public async Task
+        GetAsync_WhenRequestFromEnergySupplierWithSpecificBalanceResponsibleAndGridArea_CorrectTimeSeriesSplit()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production, TimeSeriesType.FlexConsumption },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierC,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleB);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual
+            .Select(ats => (ats.Version, ats.GridArea, ats.TimeSeriesType))
+            .Should()
+            .BeEquivalentTo([
+                (256L, "101", TimeSeriesType.FlexConsumption), (512L, "101", TimeSeriesType.FlexConsumption),
+                (1024L, "101", TimeSeriesType.FlexConsumption), (2048L, "101", TimeSeriesType.FlexConsumption),
+                (256L, "101", TimeSeriesType.Production), (512L, "101", TimeSeriesType.Production),
+                (1024L, "101", TimeSeriesType.Production), (2048L, "101", TimeSeriesType.Production)
+            ]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromEnergySupplierWithGridArea_CorrectTimeSeriesPoints()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.FlexConsumption },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeB,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierC,
+            balanceResponsibleId: null);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual.Select(ats => ats.Version).Should().BeEquivalentTo([11L, 22L, 33L, 55L, 77L, 88L]);
+
+        actual.Single(ats => ats.Version == 11L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.111]);
+
+        actual.Single(ats => ats.Version == 22L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([4.444]);
+
+        actual.Single(ats => ats.Version == 33L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.222]);
+
+        actual.Single(ats => ats.Version == 55L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.333]);
+
+        actual.Single(ats => ats.Version == 77L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.444]);
+
+        actual.Single(ats => ats.Version == 88L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([4.777]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromEnergySupplierWithGridArea_CorrectTimeSeriesSplit()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production, TimeSeriesType.FlexConsumption },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeC,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierA,
+            balanceResponsibleId: null);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual
+            .Select(ats => (ats.Version, ats.GridArea, ats.TimeSeriesType))
+            .Should()
+            .BeEquivalentTo([
+                (11L, "301", TimeSeriesType.FlexConsumption), (22L, "301", TimeSeriesType.FlexConsumption),
+                (33L, "301", TimeSeriesType.FlexConsumption), (55L, "301", TimeSeriesType.FlexConsumption),
+                (77L, "301", TimeSeriesType.FlexConsumption), (88L, "301", TimeSeriesType.FlexConsumption),
+                (11L, "301", TimeSeriesType.Production), (22L, "301", TimeSeriesType.Production),
+                (33L, "301", TimeSeriesType.Production), (55L, "301", TimeSeriesType.Production),
+                (77L, "301", TimeSeriesType.Production), (88L, "301", TimeSeriesType.Production)
+            ]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromEnergySupplierWithSpecificBalanceResponsible_CorrectTimeSeriesPoints()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.FlexConsumption },
+            gridArea: null,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierC,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleB);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        var groupedByVersionAndGridArea = actual.GroupBy(ats => (ats.Version, ats.GridArea)).ToList();
+
+        groupedByVersionAndGridArea.Select(g => g.Key)
+            .Should()
+            .BeEquivalentTo([
+                (256L, "101"), (512L, "101"), (1024L, "101"), (2048L, "101"),
+                (256L, "201"), (512L, "201"), (1024L, "201"), (2048L, "201")
+            ]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (256L, "101"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.222, 3.333]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (512L, "101"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.333]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (1024L, "101"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.444, 3.555]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (2048L, "101"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([3.666]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (256L, "201"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.111, 4.444]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (512L, "201"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.222]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (1024L, "201"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.333]);
+
+        groupedByVersionAndGridArea.Single(g => g.Key == (2048L, "201"))
+            .Single()
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([1.444, 4.777]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromEnergySupplierWithSpecificBalanceResponsible_CorrectTimeSeriesSplit()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production, TimeSeriesType.FlexConsumption },
+            gridArea: null,
+            energySupplierId: AggregatedTimeSeriesQueriesConstants.EnergySupplierC,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleB);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual
+            .Select(ats => (ats.Version, ats.GridArea, ats.TimeSeriesType))
+            .Should()
+            .BeEquivalentTo([
+                (256L, "201", TimeSeriesType.FlexConsumption),
+                (512L, "201", TimeSeriesType.FlexConsumption),
+                (1024L, "201", TimeSeriesType.FlexConsumption),
+                (2048L, "201", TimeSeriesType.FlexConsumption),
+                (256L, "201", TimeSeriesType.Production),
+                (512L, "201", TimeSeriesType.Production),
+                (1024L, "201", TimeSeriesType.Production),
+                (2048L, "201", TimeSeriesType.Production),
+                (256L, "101", TimeSeriesType.FlexConsumption),
+                (512L, "101", TimeSeriesType.FlexConsumption),
+                (1024L, "101", TimeSeriesType.FlexConsumption),
+                (2048L, "101", TimeSeriesType.FlexConsumption),
+                (256L, "101", TimeSeriesType.Production),
+                (512L, "101", TimeSeriesType.Production),
+                (1024L, "101", TimeSeriesType.Production),
+                (2048L, "101", TimeSeriesType.Production)
+            ]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromBalanceResponsibleWithGridArea_CorrectTimeSeriesPoints()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production, TimeSeriesType.FlexConsumption },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+            energySupplierId: null,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleA);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual
+            .Select(ats => (ats.Version, ats.GridArea, ats.TimeSeriesType))
+            .Should()
+            .BeEquivalentTo([
+                (1L, "101", TimeSeriesType.FlexConsumption), (2L, "101", TimeSeriesType.FlexConsumption),
+                (3L, "101", TimeSeriesType.FlexConsumption), (5L, "101", TimeSeriesType.FlexConsumption),
+                (7L, "101", TimeSeriesType.FlexConsumption), (8L, "101", TimeSeriesType.FlexConsumption),
+                (1L, "101", TimeSeriesType.Production), (2L, "101", TimeSeriesType.Production),
+                (3L, "101", TimeSeriesType.Production), (5L, "101", TimeSeriesType.Production),
+                (7L, "101", TimeSeriesType.Production), (8L, "101", TimeSeriesType.Production)
+            ]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestFromBalanceResponsibleWithGridArea_CorrectTimeSeriesSplit()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.FlexConsumption },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+            energySupplierId: null,
+            balanceResponsibleId: AggregatedTimeSeriesQueriesConstants.BalanceResponsibleB);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual.Select(ats => ats.Version).Should().BeEquivalentTo([1L, 2L, 3L, 5L, 7L, 8L]);
+
+        actual.Single(ats => ats.Version == 1L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.222, 6.666]);
+
+        actual.Single(ats => ats.Version == 2L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([4.444]);
+
+        actual.Single(ats => ats.Version == 3L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.333]);
+
+        actual.Single(ats => ats.Version == 5L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.444, 7.110]);
+
+        actual.Single(ats => ats.Version == 7L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([7.332]);
+
+        actual.Single(ats => ats.Version == 8L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([4.777]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestForGridArea_CorrectTimeSeriesPoints()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+            energySupplierId: null,
+            balanceResponsibleId: null);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual.Select(ats => ats.Version).Should().BeEquivalentTo([111L, 222L, 333L, 555L, 777L, 888L]);
+
+        actual.Single(ats => ats.Version == 111L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.222, 4.444, 6.666]);
+
+        actual.Single(ats => ats.Version == 222L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([8.888]);
+
+        actual.Single(ats => ats.Version == 333L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.444, 4.666]);
+
+        actual.Single(ats => ats.Version == 555L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.666, 4.888, 7.110]);
+
+        actual.Single(ats => ats.Version == 777L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([2.888, 7.332]);
+
+        actual.Single(ats => ats.Version == 888L)
+            .TimeSeriesPoints
+            .Select(tsp => tsp.Quantity)
+            .Should()
+            .BeEquivalentTo([9.554]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestForGridArea_CorrectTimeSeriesSplit()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[]
+            {
+                TimeSeriesType.Production, TimeSeriesType.FlexConsumption, TimeSeriesType.NetExchangePerGa,
+            },
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+            energySupplierId: null,
+            balanceResponsibleId: null);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        using var assertionScope = new AssertionScope();
+        actual
+            .Select(ats => (ats.Version, ats.GridArea, ats.TimeSeriesType))
+            .Should()
+            .BeEquivalentTo([
+                (111L, "101", TimeSeriesType.FlexConsumption), (222L, "101", TimeSeriesType.FlexConsumption),
+                (333L, "101", TimeSeriesType.FlexConsumption), (555L, "101", TimeSeriesType.FlexConsumption),
+                (777L, "101", TimeSeriesType.FlexConsumption), (888L, "101", TimeSeriesType.FlexConsumption),
+                (111L, "101", TimeSeriesType.Production), (222L, "101", TimeSeriesType.Production),
+                (333L, "101", TimeSeriesType.Production), (555L, "101", TimeSeriesType.Production),
+                (777L, "101", TimeSeriesType.Production), (888L, "101", TimeSeriesType.Production),
+                (111L, "101", TimeSeriesType.NetExchangePerGa), (222L, "101", TimeSeriesType.NetExchangePerGa),
+                (333L, "101", TimeSeriesType.NetExchangePerGa), (555L, "101", TimeSeriesType.NetExchangePerGa),
+                (777L, "101", TimeSeriesType.NetExchangePerGa), (888L, "101", TimeSeriesType.NetExchangePerGa)
+            ]);
+    }
+
+    [Fact]
+    public async Task GetAsync_WhenRequestForNoEnergySupplierBalanceResponsibleOrGridArea_CorrectTimeSeriesPoints()
+    {
+        var startOfPeriodFilter = Instant.FromUtc(2021, 12, 31, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 4, 0, 0);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: new[] { TimeSeriesType.Production },
+            gridArea: null,
+            energySupplierId: null,
+            balanceResponsibleId: null);
+
+        // Act
+        var actual = await Sut.GetAsync(parameters).ToListAsync();
+
+        var eachGridAreaIndividually = new List<AggregatedTimeSeries>();
+        foreach (var parametersForGridArea in new[]
+                 {
+                     AggregatedTimeSeriesQueriesConstants.GridAreaCodeA,
+                     AggregatedTimeSeriesQueriesConstants.GridAreaCodeB,
+                     AggregatedTimeSeriesQueriesConstants.GridAreaCodeC,
+                 }.Select(gridArea => parameters with { GridArea = gridArea }))
+        {
+            eachGridAreaIndividually.AddRange(await Sut.GetAsync(parametersForGridArea).ToListAsync());
+        }
+
+        // Assert
+        actual.Should().BeEquivalentTo(eachGridAreaIndividually);
     }
 
     [Fact]
     public async Task GetAsync_WhenRequestFromGridOperatorTotalProductionInWrongPeriod_ReturnsNoResults()
     {
         // Arrange
-        var startOfPeriodFilter = Instant.FromUtc(2020, 1, 1, 1, 1);
-        var endOfPeriodFilter = Instant.FromUtc(2021, 1, 2, 1, 1);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: GridAreaCodeA);
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: Instant.FromUtc(2020, 1, 1, 1, 1),
+            endOfPeriod: Instant.FromUtc(2021, 1, 2, 1, 1),
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA);
 
         // Act
         var actual = await Sut.GetAsync(parameters).ToListAsync();
 
         // Assert
         actual.Should().HaveCount(0);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromEnergySupplierTotalProduction_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var energySupplierIdFilter = "4321987654321";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter,
-            energySupplierId: energySupplierIdFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(FirstQuantity);
-        aggregatedTimeSeries.TimeSeriesPoints.Select(p => p.Time)
-            .ToArray()
-            .Should()
-            .AllSatisfy(p =>
-            {
-                p.Should().BeOnOrAfter(startOfPeriodFilter.ToDateTimeOffset());
-                p.Should().BeBefore(endOfPeriodFilter.ToDateTimeOffset());
-            });
     }
 
     [Fact]
     public async Task GetAsync_WhenRequestFromEnergySupplierTotalProductionBadId_ReturnsNoResults()
     {
         // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var energySupplierId = "badId";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
         var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
         var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter,
-            energySupplierId: energySupplierId);
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: [TimeSeriesType.Production],
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeC,
+            energySupplierId: "badId");
 
         // Act
         var actual = await Sut.GetAsync(parameters).ToListAsync();
@@ -191,601 +660,57 @@ public class AggregatedTimeSeriesQueriesTests : TestBase<AggregatedTimeSeriesQue
     }
 
     [Fact]
-    public async Task GetAsync_WhenRequestFromBalanceResponsibleTotalProduction_ReturnsResult()
+    public async Task GetAsync_WhenRequestReceived_PeriodFilterDenotesHalfClosedIntervalWhereStarIsIncludedAndEndIsNot()
     {
         // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var balanceResponsibleIdFilter = "1234567891234";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter,
-            balanceResponsibleId: balanceResponsibleIdFilter);
+        var startOfPeriodFilter =
+            Instant.FromDateTimeOffset(DateTimeOffset.Parse(AggregatedTimeSeriesQueriesConstants.FirstHour));
+
+        var endOfPeriodFilter =
+            Instant.FromDateTimeOffset(DateTimeOffset.Parse(AggregatedTimeSeriesQueriesConstants.SecondHour));
+
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
+
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: [TimeSeriesType.Production],
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeA);
 
         // Act
         var actual = await Sut.GetAsync(parameters).ToListAsync();
 
         // Assert
         using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
+        actual.Should().NotBeEmpty();
+        actual.Should().AllSatisfy(ats => ats.TimeSeriesPoints.Should().NotBeEmpty());
+
+        actual.SelectMany(ats => ats.TimeSeriesPoints.Select(tsp => tsp.Time))
             .Should()
-            .Equal(SecondQuantity);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromEnergySupplierPerBalanceResponsibleTotalProduction_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var balanceResponsibleIdFilter = "1234567891234";
-        var energySupplierIdFilter = "4321987654321";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter,
-            energySupplierId: energySupplierIdFilter,
-            balanceResponsibleId: balanceResponsibleIdFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(ThirdQuantity);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromEnergySupplierTotalProductionWithoutGridAreaFilter_ReturnsOneResultPerGridArea()
-    {
-        // Arrange
-        var energySupplierIdFilter = "4321987654321";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            timeSeriesType: timeSeriesTypeFilter,
-            energySupplierId: energySupplierIdFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        var resultsPerGridArea = actual.GroupBy(x => x.GridArea);
-        using var assertionScope = new AssertionScope();
-        foreach (var resultsInGridArea in resultsPerGridArea)
-        {
-            resultsInGridArea.Should().ContainSingle($"There should be only one result for grid area: {resultsInGridArea.Key}.");
-        }
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorTotalProductionFirstCorrectionSettlement_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var firstCorrectionCalculationId = Guid.NewGuid();
-        await CreateRowsForFirstCorrectionAsync(firstCorrectionCalculationId, DeltaTableCalculationType.FirstCorrectionSettlement);
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    firstCorrectionCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.CalculationType.Should().Be(CalculationType.FirstCorrectionSettlement);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(FirstQuantity, SecondQuantity);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorTotalProductionSecondCorrectionSettlement_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var secondCorrectionCalculationId = Guid.NewGuid();
-        await CreateRowsForFirstCorrectionAsync(secondCorrectionCalculationId, DeltaTableCalculationType.SecondCorrectionSettlement);
-
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    secondCorrectionCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.CalculationType.Should().Be(CalculationType.SecondCorrectionSettlement);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(FirstQuantity, SecondQuantity);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorTotalProductionThirdCorrectionSettlement_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var thirdCorrectionCalculationId = Guid.NewGuid();
-        await CreateRowsForFirstCorrectionAsync(thirdCorrectionCalculationId, DeltaTableCalculationType.ThirdCorrectionSettlement);
-
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    thirdCorrectionCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.CalculationType.Should().Be(CalculationType.ThirdCorrectionSettlement);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(FirstQuantity, SecondQuantity);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorForOneDay_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 3, 0, 0);
-
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-        var aggregatedTimeSeries = actual.First();
-        aggregatedTimeSeries!.GridArea.Should().Be(gridAreaFilter);
-        aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter);
-        aggregatedTimeSeries.TimeSeriesPoints
-            .Select(p => p.Quantity.ToString(CultureInfo.InvariantCulture))
-            .ToArray()
-            .Should()
-            .Equal(FirstQuantity);
+            .AllSatisfy(dto => dto.Should().BeOnOrAfter(startOfPeriodFilter.ToDateTimeOffset()))
+            .And
+            .AllSatisfy(dto => dto.Should().BeBefore(endOfPeriodFilter.ToDateTimeOffset()));
     }
 
     [Fact]
     public async Task GetAsync_WhenRequestFromGridOperatorStartAndEndDataAreEqual_ReturnsNoResult()
     {
         // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
+        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
+        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
 
-        await AddCreatedRowsInArbitraryOrderAsync();
+        await _aggregatedTimeSeriesQueriesData.AddDataAsync();
 
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: gridAreaFilter,
-            timeSeriesType: timeSeriesTypeFilter);
+        var parameters = AggregatedTimeSeriesQueriesData.CreateQueryParameters(
+            startOfPeriod: startOfPeriodFilter,
+            endOfPeriod: endOfPeriodFilter,
+            timeSeriesType: [TimeSeriesType.Production],
+            gridArea: AggregatedTimeSeriesQueriesConstants.GridAreaCodeC);
 
         // Act
         var actual = await Sut.GetAsync(parameters).ToListAsync();
 
         // Assert
         actual.Should().HaveCount(0);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromEnergySupplierPerBalanceResponsibleTotalProduction_ReturnsTwoResults()
-    {
-        // Arrange
-        var balanceResponsibleIdFilter = "1234567891234";
-        var energySupplierIdFilter = "4321987654321";
-        var timeSeriesTypeFilter = TimeSeriesType.Production;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            timeSeriesType: timeSeriesTypeFilter,
-            energySupplierId: energySupplierIdFilter,
-            balanceResponsibleId: balanceResponsibleIdFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(2);
-        actual.Select(result => result.GridArea).Should().Equal(GridAreaCodeB, GridAreaCodeC);
-        actual.Should().AllSatisfy(aggregatedTimeSeries => aggregatedTimeSeries.TimeSeriesType.Should().Be(timeSeriesTypeFilter));
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestFromGridOperatorTotalProductionWithEmptyGridArea_ReturnsResult()
-    {
-        // Arrange
-        var emptyGridAreaFilter = string.Empty;
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(startOfPeriodFilter, endOfPeriodFilter),
-                    _firstCalculationId,
-                    1),
-            },
-            gridArea: emptyGridAreaFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(1);
-    }
-
-    [Fact]
-    public async Task GetAsync_WithoutCalculations_ReturnsEmptyResult()
-    {
-        // Arrange
-        var startOfPeriodFilter = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var endOfPeriodFilter = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: Array.Empty<CalculationForPeriod>());
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        actual.Should().BeEmpty();
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenRequestedPeriodWithMultipleCalculations_ReturnsResult()
-    {
-        // Arrange
-        var gridAreaFilter = GridAreaCodeC;
-        var periodStartForFirstCalculation = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var periodEndForFirstCalculation = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var periodStartForSecondCalculation = Instant.FromUtc(2022, 1, 3, 0, 0);
-        var periodEndForSecondCalculation = Instant.FromUtc(2022, 1, 4, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(periodStartForFirstCalculation, periodEndForFirstCalculation),
-                    _firstCalculationId,
-                    1),
-                new CalculationForPeriod(
-                    new Period(periodStartForSecondCalculation, periodEndForSecondCalculation),
-                    _secondCalculationId,
-                    2),
-            },
-            gridArea: gridAreaFilter);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(2);
-
-        actual.Should().ContainSingle(c => c.Version == 1)
-            .Which.TimeSeriesPoints.Should().AllSatisfy(point =>
-                point.Time.Should().BeOnOrAfter(periodStartForFirstCalculation.ToDateTimeOffset()).And
-                    .BeBefore(periodEndForFirstCalculation.ToDateTimeOffset()));
-        actual.Should().ContainSingle(c => c.Version == 2)
-            .Which.TimeSeriesPoints.Should().AllSatisfy(point =>
-                point.Time.Should().BeOnOrAfter(periodStartForSecondCalculation.ToDateTimeOffset()).And
-                    .BeBefore(periodEndForSecondCalculation.ToDateTimeOffset()));
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenNoGridAreaAndMultipleCalculations_ReturnsAResultPerCalculationPerGridArea()
-    {
-        // Arrange
-        var periodStartForFirstCalculation = Instant.FromUtc(2022, 1, 1, 0, 0);
-        var periodEndForFirstCalculation = Instant.FromUtc(2022, 1, 2, 0, 0);
-        var periodStartForSecondCalculation = Instant.FromUtc(2022, 1, 3, 0, 0);
-        var periodEndForSecondCalculation = Instant.FromUtc(2022, 1, 4, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync();
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(periodStartForFirstCalculation, periodEndForFirstCalculation),
-                    _firstCalculationId,
-                    1),
-                new CalculationForPeriod(
-                    new Period(periodStartForSecondCalculation, periodEndForSecondCalculation),
-                    _secondCalculationId,
-                    2),
-            });
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        using var assertionScope = new AssertionScope();
-        actual.Should().HaveCount(3);
-        actual.Should().ContainSingle(x => x.GridArea == GridAreaCodeC && x.Version == 1);
-        actual.Should().ContainSingle(x => x.GridArea == GridAreaCodeC && x.Version == 2);
-        actual.Should().ContainSingle(x => x.GridArea == GridAreaCodeA && x.Version == 2);
-    }
-
-    [Fact]
-    public async Task GetAsync_WhenNoResultForRequestedCalculation_ReturnsEmptyResult()
-    {
-        // Arrange
-        var calculationWithoutAnyResults = Guid.NewGuid();
-        var periodStart = Instant.FromUtc(2024, 1, 1, 23, 0, 0);
-        var periodEnd = Instant.FromUtc(2024, 1, 3, 23, 0, 0);
-        await AddCreatedRowsInArbitraryOrderAsync(addSecondCorrection: true);
-        var parameters = CreateQueryParameters(
-            latestCalculationForPeriods: new[]
-            {
-                new CalculationForPeriod(
-                    new Period(periodStart, periodEnd),
-                    calculationWithoutAnyResults,
-                    1),
-            },
-            gridArea: GridAreaCodeC);
-
-        // Act
-        var actual = await Sut.GetAsync(parameters).ToListAsync();
-
-        // Assert
-        actual.Should().BeEmpty();
-    }
-
-    private AggregatedTimeSeriesQueryParameters CreateQueryParameters(
-        IReadOnlyCollection<CalculationForPeriod> latestCalculationForPeriods,
-        TimeSeriesType? timeSeriesType = null,
-        string? gridArea = null,
-        string? energySupplierId = null,
-        string? balanceResponsibleId = null)
-    {
-        return new AggregatedTimeSeriesQueryParameters(
-            TimeSeriesTypes: new[] { timeSeriesType ?? TimeSeriesType.Production },
-            GridArea: gridArea,
-            EnergySupplierId: energySupplierId,
-            BalanceResponsibleId: balanceResponsibleId,
-            LatestCalculationForPeriod: latestCalculationForPeriods);
-    }
-
-    private async Task AddCreatedRowsInArbitraryOrderAsync(
-        bool addFirstCorrection = false,
-        bool addSecondCorrection = false,
-        bool addThirdCorrection = false)
-    {
-        const string firstCalculationResultId = "aaaaaaaa-386f-49eb-8b56-63fae62e4fc7";
-        const string secondCalculationResultId = "bbbbbbbb-b58b-4190-a873-eded0ed50c20";
-        const string thirdCalculationResultId = "cccccccc-386f-49eb-8b56-63fae62e4fc7";
-
-        const string firstHour = "2022-01-01T01:00:00.000Z";
-        const string secondHour = "2022-01-01T02:00:00.000Z";
-        const string thirdHour = "2022-01-01T03:00:00.000Z";
-
-        const string secondDay = "2022-01-02T00:00:00.000Z";
-        const string thirdDay = "2022-01-03T00:00:00.000Z";
-
-        const string energySupplier = "4321987654321";
-        const string balanceResponsibleId = "1234567891234";
-        var row1 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantity);
-        var row2 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _secondCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantity);
-
-        var row3 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: secondCalculationResultId, time: secondHour, gridArea: GridAreaCodeC, quantity: ThirdQuantity, calculationExecutionTimeStart: "2022-03-12T03:00:00.000Z");
-        var row4 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: secondCalculationResultId, time: thirdHour, gridArea: GridAreaCodeC, quantity: FourthQuantity, calculationExecutionTimeStart: "2022-03-12T03:00:00.000Z");
-
-        var row5 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndGridArea, energySupplierId: energySupplier);
-        var row6 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: secondCalculationResultId, time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantity, aggregationLevel: DeltaTableAggregationLevel.BalanceResponsibleAndGridArea, balanceResponsibleId: balanceResponsibleId);
-
-        var row7 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: secondHour, gridArea: GridAreaCodeC, quantity: ThirdQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndBalanceResponsibleAndGridArea, balanceResponsibleId: balanceResponsibleId, energySupplierId: energySupplier);
-        var row8 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: thirdHour, gridArea: GridAreaCodeB, quantity: FourthQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndBalanceResponsibleAndGridArea, balanceResponsibleId: balanceResponsibleId, energySupplierId: energySupplier);
-
-        var row9 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: firstHour, gridArea: GridAreaCodeA, quantity: FirstQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndBalanceResponsibleAndGridArea);
-        var row10 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _secondCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: thirdHour, gridArea: GridAreaCodeB, quantity: FourthQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndBalanceResponsibleAndGridArea, balanceResponsibleId: balanceResponsibleId, energySupplierId: energySupplier);
-
-        var row11 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: firstHour, gridArea: GridAreaCodeB, quantity: FirstQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndGridArea, energySupplierId: energySupplier);
-        var row12 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: secondHour, gridArea: GridAreaCodeB, quantity: FirstQuantity, aggregationLevel: DeltaTableAggregationLevel.EnergySupplierAndGridArea, energySupplierId: energySupplier);
-
-        var row1FirstCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, calculationType: DeltaTableCalculationType.FirstCorrectionSettlement, time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantityFirstCorrection);
-        var row2FirstCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, calculationType: DeltaTableCalculationType.FirstCorrectionSettlement, time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantityFirstCorrection);
-
-        var row1SecondCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: secondCalculationResultId, calculationType: DeltaTableCalculationType.SecondCorrectionSettlement, time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantitySecondCorrection);
-        var row2SecondCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: secondCalculationResultId, calculationType: DeltaTableCalculationType.SecondCorrectionSettlement, time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantitySecondCorrection);
-
-        var row1ThirdCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: thirdCalculationResultId, calculationType: DeltaTableCalculationType.ThirdCorrectionSettlement, time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantityThirdCorrection);
-        var row2ThirdCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: thirdCalculationResultId, calculationType: DeltaTableCalculationType.ThirdCorrectionSettlement, time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantityThirdCorrection);
-        var row4ThirdCorrection = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: thirdCalculationResultId, calculationType: DeltaTableCalculationType.ThirdCorrectionSettlement, time: thirdHour, gridArea: GridAreaCodeC, quantity: FourthQuantityThirdCorrection);
-
-        var row1SecondDay = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: secondDay, gridArea: GridAreaCodeC, quantity: FirstQuantity);
-        var row1ThirdDay = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _firstCalculationId.ToString(), calculationResultId: firstCalculationResultId, time: thirdDay, gridArea: GridAreaCodeC, quantity: SecondQuantity);
-
-        var row1SecondBatchThirdDay = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _secondCalculationId.ToString(), calculationResultId: secondCalculationResultId, time: thirdDay, gridArea: GridAreaCodeC, quantity: SecondQuantity);
-        var row2SecondBatchThirdDay = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: _secondCalculationId.ToString(), calculationResultId: secondCalculationResultId, time: thirdDay, gridArea: GridAreaCodeA, quantity: SecondQuantity);
-
-        var rows = new List<IReadOnlyCollection<string>>
-        {
-            row1,
-            row2,
-            row3,
-            row4,
-            row5,
-            row6,
-            row7,
-            row8,
-            row9,
-            row10,
-            row11,
-            row12,
-            row1SecondDay,
-            row1ThirdDay,
-            row1SecondBatchThirdDay,
-            row2SecondBatchThirdDay,
-        };
-
-        if (addFirstCorrection)
-        {
-            rows.AddRange(new[]
-            {
-                row1FirstCorrection,
-                row2FirstCorrection,
-            });
-        }
-
-        if (addSecondCorrection)
-        {
-            rows.AddRange(new[]
-            {
-                row1SecondCorrection,
-                row2SecondCorrection,
-            });
-        }
-
-        if (addThirdCorrection)
-        {
-            rows.AddRange(new[]
-            {
-                row1ThirdCorrection,
-                row2ThirdCorrection,
-                row4ThirdCorrection,
-            });
-        }
-
-        rows = rows.OrderBy(_ => Guid.NewGuid()).ToList();
-
-        await _fixture.DatabricksSchemaManager.EmptyAsync(_fixture.DatabricksSchemaManager.DeltaTableOptions.Value.ENERGY_RESULTS_TABLE_NAME);
-        await _fixture.DatabricksSchemaManager.InsertAsync<EnergyResultColumnNames>(_fixture.DatabricksSchemaManager.DeltaTableOptions.Value.ENERGY_RESULTS_TABLE_NAME, rows);
-    }
-
-    private async Task CreateRowsForFirstCorrectionAsync(Guid calculationId, string calculationType)
-    {
-        const string firstHour = "2022-01-01T01:00:00.000Z";
-        const string secondHour = "2022-01-01T02:00:00.000Z";
-
-        var row1 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: calculationId.ToString(), calculationResultId: Guid.NewGuid().ToString(), time: firstHour, gridArea: GridAreaCodeC, quantity: FirstQuantity, calculationType: calculationType);
-        var row2 = EnergyResultDeltaTableHelper.CreateRowValues(calculationId: calculationId.ToString(), calculationResultId: Guid.NewGuid().ToString(), time: secondHour, gridArea: GridAreaCodeC, quantity: SecondQuantity, calculationType: calculationType);
-        var rows = new List<IReadOnlyCollection<string>> { row1, row2, };
-
-        rows = rows.OrderBy(_ => Guid.NewGuid()).ToList();
-        await _fixture.DatabricksSchemaManager.EmptyAsync(_fixture.DatabricksSchemaManager.DeltaTableOptions.Value.ENERGY_RESULTS_TABLE_NAME);
-        await _fixture.DatabricksSchemaManager.InsertAsync<EnergyResultColumnNames>(_fixture.DatabricksSchemaManager.DeltaTableOptions.Value.ENERGY_RESULTS_TABLE_NAME, rows);
     }
 }
